+++ test/jasmine/spec/Core_beforeKeyDownSpec.js
@@ -34,7 +34,7 @@ describe('Core_beforeKeyDown', function () {
     });
   });
 
-  it('should prevent hook fron runing default action', function () {
+  it('should prevent hook fron running default action', function () {
     var called = false;
 
     handsontable({
@@ -85,6 +85,25 @@ describe('Core_beforeKeyDown', function () {
     runs(function () {
       expect(getData().length).toEqual(3);
       expect(getSelected()).toEqual([0,1,0,1]);
+    });
+  });
+
+  it('should run beforeKeyDown hook in cell editor handler', function () {
+    var called = 0;
+
+    handsontable({
+      data : [[1,2,3,4,5],[1,2,3,4,5]],
+      beforeKeyDown: function (event) {
+        called++;
+      }
+    });
+    selectCell(0,0);
+
+    keyDown('enter');
+    keyDown('enter');
+
+    runs(function () {
+      expect(called).toEqual(2);
     });
   });